requirements.txt
asgiref==3.8.1
Django==3.2
djangorestframework==3.15.1
lxml==5.2.1
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
